changes and tests around plugin close
